requirements.txt
awscli==1.16.152
boto3==1.9.152
djangorestframework==3.10.3
django==2.2.5
django-allauth==0.40.0
django-cors-headers==3.1.0
django-extensions==2.1.6
django-filter==2.1.0
django-ranged-fileresponse==0.1.2
django-rest-auth==0.9.5
django-storages==1.7.1
django-environ==0.4.5
future==0.16.0
gunicorn==19.9.0
jmespath==0.9.3
pinax-badges==2.0.0
psycopg2-binary==2.8.2
python-dateutil==2.6.1
sentry-sdk==0.11.2
whoosh==2.7.4
zappa==0.47.0